Update bm_NukeColourConverter.py
Fix Nuke 13 compatability

+++ python/bm_NukeColourConverter.py
@@ -1,10 +1,10 @@
 # --------------------------------------------------------------
 #  bm_NukeColourConverter.py
-#  Version: 1.0.0
+#  Version: 1.1.0
 #  Author: Ben McEwan
 #
 #  Last Modified by: Ben McEwan
-#  Last Updated: June 24th, 2019
+#  Last Updated: March 7th, 2022
 # --------------------------------------------------------------
 
 # --------------------------------------------------------------
@@ -20,4 +20,4 @@ def hex_colour_to_int(hexValue):
     return int(hexValue+'00', 16)
 
 def hex_colour_to_rgb(red, green, blue):
-    return int('%02x%02x%02x%02x' % (red*255, green*255, blue*255, 255),16)
+    return int('%02x%02x%02x%02x' % (int(red*255), int(green*255), int(blue*255), 255),16)
